completed get file tree interface

+++ backend/app/api/file.py
@@ -1,9 +1,10 @@
 from flask import Blueprint, send_file
 from flask_restful import Api, Resource, abort, reqparse
 from werkzeug.datastructures import FileStorage
+from flask_login import login_required
 
 from app.service.file import FileService
-
+from app.checker import check_project_permission
 bp = Blueprint(
     "file",
     __name__
@@ -18,21 +19,30 @@ parser.add_argument('file', type=FileStorage, location='files')
 
 
 class File(Resource):
+    @login_required
     def get(self, project_id, path):
+        if not check_project_permission(project_id, 'read'):
+            abort(404, message="Permission denied")
         res, flag = file_service.download_file(path, project_id)
         if flag:
             return send_file(res)
         else:
             abort(404, message=res)
 
+    @login_required
     def post(self, project_id, path):
+        if not check_project_permission(project_id, 'write'):
+            abort(404, message="Permission denied")
         res, flag = file_service.create_file(path, project_id)
         if flag:
             return '', 204
         else:
             abort(400, message=res)
 
+    @login_required
     def put(self, project_id, path):
+        if not check_project_permission(project_id, 'write'):
+            abort(404, message="Permission denied")
         args = parser.parse_args()
         file = args['file']
         if not file:
@@ -43,7 +53,10 @@ class File(Resource):
         else:
             abort(400, message=res)
 
+    @login_required
     def delete(self, project_id, path):
+        if not check_project_permission(project_id, 'write'):
+            abort(404, message="Permission denied")
         res, flag = file_service.remove_file(path, project_id)
         if flag:
             return "", 204
